新しいファイル管理

+++ src/main/java/jp/scid/bio/store/folder/FolderRecordBasicFolder.java
@@ -4,32 +4,50 @@ import java.io.File;
 import java.io.IOException;
 import java.text.ParseException;
 
-import jp.scid.bio.store.folder.FolderRecordGroupFolder.Source;
 import jp.scid.bio.store.jooq.tables.records.FolderRecord;
 import jp.scid.bio.store.sequence.FolderContentGeneticSequence;
 import jp.scid.bio.store.sequence.GeneticSequence;
+import jp.scid.bio.store.sequence.ImportableSequenceSource;
 
-public class FolderRecordBasicFolder extends AbstractFolder {
-    public FolderRecordBasicFolder(FolderRecord record, Source folderSource) {
+public class FolderRecordBasicFolder extends AbstractFolder implements ImportableSequenceSource {
+    public FolderRecordBasicFolder(FolderRecord record, AbstractFolder.Source folderSource) {
         super(record, folderSource);
     }
 
+    public FolderContentGeneticSequence createContent(GeneticSequence sequence) {
+        return source.createFolderContent(sequence, this);
+    }
+    
     public FolderContentGeneticSequence addSequence(GeneticSequence sequence) {
-        FolderContentGeneticSequence content = source.createFolderContent(sequence, this);
+        FolderContentGeneticSequence content = createContents(sequence);
         content.save();
-        sequences.add(content);
+        
+        fireSequencesChange();
         return content;
     }
     
     public FolderContentGeneticSequence importSequence(File file) throws IOException, ParseException {
         GeneticSequence sequence = source.createGeneticSequence(file);
-        sequence.save();
-        return addSequence(sequence);
+        FolderContentGeneticSequence content = createContents(sequence);
+        content.save();
+        
+        fireSequencesChange();
+        return content;
     }
     
-    public FolderContentGeneticSequence removeSequenceAt(int index) {
-        FolderContentGeneticSequence sequence = sequences.removeElementAt(index);
-        sequence.delete();
-        return sequence;
+    private FolderContentGeneticSequence createContents(GeneticSequence sequence) {
+        return source.createFolderContent(sequence, this);
+    }
+    
+    public boolean removeSequence(FolderContentGeneticSequence sequence) {
+        if (sequence.folderId() == null || !id().equals(sequence.folderId())) {
+            return false;
+        }
+        
+        boolean result = sequence.delete();
+        if (result) {
+            fireSequencesChange();
+        }
+        return result;
     }
 }
